pyproject.toml
[project]
name = "mdpub"
version = "0.1.0"
description = "A CLI-first app to decompose markdown documents into a structured format"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "kotulc" }]
dependencies = [
  "typer>=0.12",
  "pydantic>=2.7",
  "PyYAML>=6.0",
  "sqlmodel>=0.0.21",
  "SQLAlchemy>=2.0",
]

[project.optional-dependencies]
postgres = [
  "psycopg[binary]>=3.2",
]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "mypy>=1.10",
]

[project.scripts]
mdpub = "mdpub.cli:app"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mdpub"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = ["src"]
